LayoutTests/webgpu/webgpu-enabled-in-worker.html

<!DOCTYPE html><!-- webkit-test-runner [ experimental:WebGPUEnabled=true ] -->
<html>
<meta charset=utf-8>
<title>Test if WebGPU is enabled for workers.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
if (window.testRunner)
    testRunner.waitUntilDone();

const worker = new Worker('js/create-worker-device.js');

worker.onmessage = event => {
    test(() => { assert_true(event.data, "Create GPUDevice with a worker."); });

    if (window.testRunner)
        testRunner.notifyDone();
};
</script>
</html>
